MCP_Aug_07-2025/MCP/requirements.txt
Package                   Version
------------------------- -----------
aiohappyeyeballs          2.6.1
aiohttp                   3.12.15
aiosignal                 1.4.0
altair                    5.5.0
annotated-types           0.7.0
anyio                     4.10.0
attrs                     25.3.0
Authlib                   1.6.1
blinker                   1.9.0
cachetools                5.5.2
certifi                   2025.8.3
cffi                      1.17.1
charset-normalizer        3.4.2
click                     8.2.1
cryptography              45.0.6
cyclopts                  3.22.5
dataclasses-json          0.6.7
distro                    1.9.0
dnspython                 2.7.0
docstring_parser          0.17.0
docutils                  0.22
email_validator           2.2.0
exceptiongroup            1.3.0
fastmcp                   2.11.1
frozenlist                1.7.0
gitdb                     4.0.12
GitPython                 3.1.45
google-api-core           2.25.1
google-api-python-client  2.178.0
google-auth               2.40.3
google-auth-httplib2      0.2.0
google-auth-oauthlib      1.2.2
googleapis-common-protos  1.70.0
h11                       0.16.0
httpcore                  1.0.9
httplib2                  0.22.0
httpx                     0.28.1
httpx-sse                 0.4.1
idna                      3.10
isodate                   0.7.2
Jinja2                    3.1.6
jiter                     0.10.0
jsonpatch                 1.33
jsonpointer               3.0.0
jsonschema                4.25.0
jsonschema-path           0.3.4
jsonschema-specifications 2025.4.1
langchain                 0.3.27
langchain-community       0.3.27
langchain-core            0.3.72
langchain-mcp-adapters    0.1.9
langchain-openai          0.3.28
langchain-text-splitters  0.3.9
langgraph                 0.6.3
langgraph-checkpoint      2.1.1
langgraph-prebuilt        0.6.3
langgraph-sdk             0.2.0
langsmith                 0.4.12
lazy-object-proxy         1.11.0
markdown-it-py            3.0.0
MarkupSafe                3.0.2
marshmallow               3.26.1
mcp                       1.12.3
mdurl                     0.1.2
more-itertools            10.7.0
multidict                 6.6.3
mypy_extensions           1.1.0
narwhals                  2.0.1
numpy                     2.3.2
oauthlib                  3.3.1
openai                    1.99.1
openapi-core              0.19.5
openapi-pydantic          0.5.1
openapi-schema-validator  0.6.3
openapi-spec-validator    0.7.2
orjson                    3.11.1
ormsgpack                 1.10.0
packaging                 25.0
pandas                    2.3.1
parse                     1.20.2
pathable                  0.4.4
pillow                    11.3.0
pip                       25.2
propcache                 0.3.2
proto-plus                1.26.1
protobuf                  6.31.1
pyarrow                   21.0.0
pyasn1                    0.6.1
pyasn1_modules            0.4.2
pycparser                 2.22
pydantic                  2.11.7
pydantic_core             2.33.2
pydantic-settings         2.10.1
pydeck                    0.9.1
Pygments                  2.19.2
pyparsing                 3.2.3
pyperclip                 1.9.0
python-dateutil           2.9.0.post0
python-dotenv             1.1.1
python-multipart          0.0.20
pytz                      2025.2
PyYAML                    6.0.2
referencing               0.36.2
regex                     2025.7.34
requests                  2.32.4
requests-oauthlib         2.0.0
requests-toolbelt         1.0.0
rfc3339-validator         0.1.4
rich                      14.1.0
rich-rst                  1.3.1
rpds-py                   0.26.0
rsa                       4.9.1
setuptools                65.5.0
six                       1.17.0
smmap                     5.0.2
sniffio                   1.3.1
SQLAlchemy                2.0.42
sse-starlette             3.0.2
starlette                 0.47.2
streamlit                 1.48.0
tenacity                  9.1.2
tiktoken                  0.10.0
toml                      0.10.2
tornado                   6.5.1
tqdm                      4.67.1
typing_extensions         4.14.1
typing-inspect            0.9.0
typing-inspection         0.4.1
tzdata                    2025.2
uritemplate               4.2.0
urllib3                   2.5.0
uvicorn                   0.35.0
Werkzeug                  3.1.1
xxhash                    3.5.0
yarl                      1.20.1
zstandard                 0.23.0
